+++ java/src/main/java/io/github/flanglet/kanzi/entropy/ANSRangeEncoder.java
@@ -123,7 +123,7 @@ public class ANSRangeEncoder implements EntropyEncoder {
     this.symbols = new Symbol[dim][256];
     this.buffer = new byte[0];
     this.logRange = (order == 0) ? logRange : Math.max(logRange - 1, 8);
-    this.chunkSize = Math.min(chunkSize << (8 * order), MAX_CHUNK_SIZE);
+    this.chunkSize = (int) Math.min(((long) chunkSize) << (8 * order), MAX_CHUNK_SIZE);
 
     for (int i = 0; i < dim; i++) {
       this.freqs[i] = new int[257];
@@ -152,7 +152,8 @@ public class ANSRangeEncoder implements EntropyEncoder {
     this.symbols = new Symbol[dim][256];
     this.buffer = new byte[0];
     this.logRange = (order == 0) ? DEFAULT_LOG_RANGE : DEFAULT_LOG_RANGE - 1;
-    this.chunkSize = Math.min(DEFAULT_ANS0_CHUNK_SIZE << (8 * order), MAX_CHUNK_SIZE);
+    this.chunkSize =
+        (int) Math.min(((long) DEFAULT_ANS0_CHUNK_SIZE) << (8 * order), MAX_CHUNK_SIZE);
 
     for (int i = 0; i < dim; i++) {
       this.freqs[i] = new int[257];
